Calculates number of complete observations per monitor
Calculates number of complete observations per monitor using function
call complete(directory, ID) where directory is the subdirectory (of the
R working directory) containing the monitor data files, and the ID is
the id number or range of numbers for the monitor(s) of interest. Has
some odd bits to deal with decreasing ID ranges.